% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_function.R
\name{hello_function}
\alias{hello_function}
\title{A Hello Function}
\usage{
hello_function(text)
}
\arguments{
\item{text}{To whom you want to say hello.}
}
\description{
This function allows you to say hello.
}
\examples{
hello_function()
}
\keyword{hello}
