% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senti.R
\name{senti}
\alias{senti}
\title{sentiment analysis}
\usage{
senti(
  edgelist,
  target,
  positiveLexicon = system.file("extdata", "SeedsPos.txt", package = "senet"),
  negativeLexicon = system.file("extdata", "SeedsNeg.txt", package = "senet")
)
}
\arguments{
\item{target}{the target keyword string}

\item{positiveLexicon}{the positive keyword string list}

\item{negativeLexicon}{the negative keyword string list}

\item{graph}{edge graph}
}
\description{
sentiment analysis
}
