% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tograph.R
\name{tograph}
\alias{tograph}
\title{SENET package in R
This package reads text files and generates measures of text sentiments based on shortest path in the semantic network}
\usage{
tograph(
  filename,
  window_size = 3,
  useDroplist = TRUE,
  droplist = system.file("extdata", "droplist.txt", package = "senet")
)
}
\arguments{
\item{filename}{path to the text file}

\item{window_size}{size of the n-gram, defult is 3}

\item{useDroplist}{whether to drop a list of words, default is yes}

\item{droplist}{path to the stopword list, if not specified, the built-in stopword list is used}
}
\description{
SENET package in R
This package reads text files and generates measures of text sentiments based on shortest path in the semantic network
}
