% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_function.R
\name{hello_function}
\alias{hello_function}
\title{A Hello Function}
\usage{
hello_function(filename)
}
\arguments{
\item{text_file_path}{file path to the text doc}
}
\description{
This function allows you to say hello.
}
\examples{
hello_function()
}
\keyword{hello}
