% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senti.R
\name{senti}
\alias{senti}
\title{sentiment analysis}
\usage{
senti(
  edgelist,
  target,
  positiveLexicon = system.file("extdata", "SeedsPos.txt", package = "senet"),
  negativeLexicon = system.file("extdata", "SeedsNeg.txt", package = "senet"),
  verbose = FALSE
)
}
\arguments{
\item{edgelist}{edgelist with edges and counts}

\item{target}{the target keyword string}

\item{positiveLexicon}{path to the positive sentiment words in txt format, if not specified, the built-in lexicon is used}

\item{negativeLexicon}{path to the negtive sentiment words in txt format, if not specified, the built-in lexicon is used}

\item{verbose}{print messages}
}
\description{
sentiment analysis
}
